Split TrustChain implementation
