+++ Tribler/Test/Community/Triblerchain/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains tests for the TriblerChain.
+"""
