+++ Tribler/Test/Community/Trustchain/test_block.py
@@ -8,14 +8,59 @@ from Tribler.community.trustchain.block import (TrustChainBlock, GENESIS_HASH, E
 from Tribler.Test.Community.Trustchain.test_trustchain_utilities import TrustChainTestCase, TestBlock
 
 
+class MockDatabase(object):
+    """
+    This mocked database is only used during the tests.
+    """
+
+    def __init__(self):
+        super(MockDatabase, self).__init__()
+        self.data = dict()
+
+    def add_block(self, block):
+        if self.data.get(block.public_key) is None:
+            self.data[block.public_key] = []
+        self.data[block.public_key].append(block)
+        self.data[block.public_key].sort(key=lambda b: b.sequence_number)
+
+    def get(self, pk, seq):
+        if self.data.get(pk) is None:
+            return None
+        item = [i for i in self.data[pk] if i.sequence_number == seq]
+        return item[0] if item else None
+
+    def get_linked(self, blk):
+        if self.data.get(blk.link_public_key) is None:
+            return None
+        item = [i for i in self.data[blk.link_public_key] if
+                i.sequence_number == blk.link_sequence_number or i.link_sequence_number == blk.sequence_number]
+        return item[0] if item else None
+
+    def get_latest(self, pk):
+        return self.data[pk][-1] if self.data.get(pk) else None
+
+    def get_block_after(self, blk):
+        if self.data.get(blk.public_key) is None:
+            return None
+        item = [i for i in self.data[blk.public_key] if i.sequence_number > blk.sequence_number]
+        return item[0] if item else None
+
+    def get_block_before(self, blk):
+        if self.data.get(blk.public_key) is None:
+            return None
+        item = [i for i in self.data[blk.public_key] if i.sequence_number < blk.sequence_number]
+        return item[-1] if item else None
+
+
 class TestBlocks(TrustChainTestCase):
-    def __init__(self, *args, **kwargs):
-        super(TestBlocks, self).__init__(*args, **kwargs)
+    """
+    This class contains tests for a TrustChain block.
+    """
 
     def test_hash(self):
         block = TrustChainBlock()
-        self.assertEqual(block.hash, "\xa1c!\x14\x11\x14\xe4\xb1g\xebB\xae\xc1y-\x0eF\x1d\x94'\x1co\xc5\xe4g\x80\xf1"
-                                     "\xc1z\xb0\x12\xd7")
+        self.assertEqual(block.hash, '\x1f\x1bp\x90\xe3>\x83\xf6\xcd\xafd\xd9\xee\xfb"&|<ZLsyB:Z\r'
+                                     '<\xc5\xb0\x97\xa3\xaf')
 
     def test_sign(self):
         crypto = ECCrypto()
@@ -24,29 +69,29 @@ class TestBlocks(TrustChainTestCase):
 
     def test_create_genesis(self):
         key = ECCrypto().generate_key(u"curve25519")
-        db = self.MockDatabase()
-        block = TrustChainBlock.create(db, key.pub().key_to_bin(), link=None)
+        db = MockDatabase()
+        block = TrustChainBlock.create({'id': 42}, db, key.pub().key_to_bin(), link=None)
         self.assertEqual(block.previous_hash, GENESIS_HASH)
         self.assertEqual(block.sequence_number, GENESIS_SEQ)
         self.assertEqual(block.public_key, key.pub().key_to_bin())
         self.assertEqual(block.signature, EMPTY_SIG)
 
     def test_create_next(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         prev = TestBlock()
         prev.sequence_number = GENESIS_SEQ
         db.add_block(prev)
-        block = TrustChainBlock.create(db, prev.public_key, link=None)
+        block = TrustChainBlock.create({'id': 42}, db, prev.public_key, link=None)
         self.assertEqual(block.previous_hash, prev.hash)
         self.assertEqual(block.sequence_number, 2)
         self.assertEqual(block.public_key, prev.public_key)
 
     def test_create_link_genesis(self):
         key = ECCrypto().generate_key(u"curve25519")
-        db = self.MockDatabase()
+        db = MockDatabase()
         link = TestBlock()
         db.add_block(link)
-        block = TrustChainBlock.create(db, key.pub().key_to_bin(), link=link)
+        block = TrustChainBlock.create({'id': 42}, db, key.pub().key_to_bin(), link=link)
         self.assertEqual(block.previous_hash, GENESIS_HASH)
         self.assertEqual(block.sequence_number, GENESIS_SEQ)
         self.assertEqual(block.public_key, key.pub().key_to_bin())
@@ -54,13 +99,13 @@ class TestBlocks(TrustChainTestCase):
         self.assertEqual(block.link_sequence_number, link.sequence_number)
 
     def test_create_link_next(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         prev = TestBlock()
         prev.sequence_number = GENESIS_SEQ
         db.add_block(prev)
         link = TestBlock()
         db.add_block(link)
-        block = TrustChainBlock.create(db, prev.public_key, link=link)
+        block = TrustChainBlock.create({'id': 42}, db, prev.public_key, link=link)
         self.assertEqual(block.previous_hash, prev.hash)
         self.assertEqual(block.sequence_number, 2)
         self.assertEqual(block.public_key, prev.public_key)
@@ -69,30 +114,24 @@ class TestBlocks(TrustChainTestCase):
 
     def test_pack(self):
         block = TrustChainBlock()
-        block.up = 3251690667711950702
-        block.down = 7431046511915463784
-        block.total_up = 7020667011326177138
-        block.total_down = 2333265293611395173
+        block.transaction = {'id': 3}
         block.public_key = ' fish, so sad that it should come to this. - We tried to warn you all but '
         block.sequence_number = 1869095012
         block.link_public_key = 'ear! - You may not share our intellect, which might explain your disrespec'
         block.link_sequence_number = 1949048934
         block.previous_hash = 'or all the natural wonders that '
         block.signature = 'grow around you. - So long, so long, and thanks for all the fish'
-        self.assertEqual(block.pack(), '- So long and thanks for all the fish, so sad that it should come to this. - We'
-                                       ' tried to warn you all but oh dear! - You may not share our intellect, which '
-                                       'might explain your disrespect, for all the natural wonders that grow around you'
-                                       '. - So long, so long, and thanks for all the fish')
+        self.assertEqual(block.pack(), ' fish, so sad that it should come to this. - We tried to warn you '
+                                       'all but oh dear! - You may not share our intellect, which might explain your '
+                                       'disrespect, for all the natural wonders that grow around you. - So long, '
+                                       'so long, and thanks for all the fish\x00\x00\x00\na1d2bid1i3')
 
     def test_unpack(self):
-        block = TrustChainBlock.unpack('- So long and thanks for all the fish, so sad that it should come to this. - '
-                                       'We tried to warn you all but oh dear! - You may not share our intellect, which '
-                                       'might explain your disrespect, for all the natural wonders that grow around '
-                                       'you. - So long, so long, and thanks for all the fish')
-        self.assertEqual(block.up, 3251690667711950702)
-        self.assertEqual(block.down, 7431046511915463784)
-        self.assertEqual(block.total_up, 7020667011326177138)
-        self.assertEqual(block.total_down, 2333265293611395173)
+        block = TrustChainBlock.unpack(' fish, so sad that it should come to this. - We tried to warn you '
+                                       'all but oh dear! - You may not share our intellect, which might explain your '
+                                       'disrespect, for all the natural wonders that grow around you. - So long, '
+                                       'so long, and thanks for all the fish\x00\x00\x00\na1d2bid1i3')
+        self.assertEqual(block.transaction, {'id': 3})
         self.assertEqual(block.public_key, ' fish, so sad that it should come to this. - We tried to warn you all but ')
         self.assertEqual(block.sequence_number, 1869095012)
         self.assertEqual(block.link_public_key,
@@ -103,7 +142,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_existing(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, block3, _) = TestBlocks.setup_validate()
         db.add_block(block1)
         db.add_block(block2)
@@ -115,7 +154,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_non_existing(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, block3, _) = TestBlocks.setup_validate()
         db.add_block(block1)
         db.add_block(block3)
@@ -126,7 +165,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_no_info(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (_, _, _, block4) = TestBlocks.setup_validate()
         db.add_block(block4)
         # Act
@@ -136,7 +175,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_partial_prev(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (_, block2, block3, _) = TestBlocks.setup_validate()
         db.add_block(block2)
         db.add_block(block3)
@@ -147,7 +186,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_partial_next(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (_, block2, block3, _) = TestBlocks.setup_validate()
         db.add_block(block2)
         db.add_block(block3)
@@ -158,7 +197,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_partial_prev_with_gap(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, block3, _) = TestBlocks.setup_validate()
         block2.previous_hash = block2.hash
         block2.sequence_number += 1
@@ -174,7 +213,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_partial_next_with_gap(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, block3, _) = TestBlocks.setup_validate()
         block3.sequence_number += 1
         db.add_block(block1)
@@ -186,7 +225,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_partial_left_gap(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, block3, _) = TestBlocks.setup_validate()
         db.add_block(block1)
         db.add_block(block3)
@@ -197,7 +236,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_partial_right_gap_from_genesis(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, block3, _) = TestBlocks.setup_validate()
         db.add_block(block1)
         db.add_block(block3)
@@ -208,7 +247,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_partial_right_gap(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, block3, _) = TestBlocks.setup_validate()
         block1.previous_hash = block3.previous_hash
         block1.sequence_number += 1
@@ -222,7 +261,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_partial_with_both_gaps(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, block3, _) = TestBlocks.setup_validate()
         block2.previous_hash = block2.hash
         block2.sequence_number += 1
@@ -235,73 +274,9 @@ class TestBlocks(TrustChainTestCase):
         # Assert
         self.assertEqual(result[0], ValidationResult.partial)
 
-    def test_validate_existing_up(self):
-        # Arrange
-        db = self.MockDatabase()
-        (block1, block2, block3, _) = TestBlocks.setup_validate()
-        db.add_block(block1)
-        db.add_block(block2)
-        db.add_block(block3)
-        # Act
-        block2 = TrustChainBlock(block2.pack_db_insert())
-        block2.up += 10
-        block2.sign(db.get(block2.public_key, block2.sequence_number).key)
-        result = block2.validate(db)
-        # Assert
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn('Total up is lower than expected compared to the preceding block', result[1])
-
-    def test_validate_existing_down(self):
-        # Arrange
-        db = self.MockDatabase()
-        (block1, block2, block3, _) = TestBlocks.setup_validate()
-        db.add_block(block1)
-        db.add_block(block2)
-        db.add_block(block3)
-        # Act
-        block2 = TrustChainBlock(block2.pack_db_insert())
-        block2.down += 10
-        block2.sign(db.get(block2.public_key, block2.sequence_number).key)
-        result = block2.validate(db)
-        # Assert
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn('Total down is lower than expected compared to the preceding block', result[1])
-
-    def test_validate_existing_total_up(self):
-        # Arrange
-        db = self.MockDatabase()
-        (block1, block2, block3, _) = TestBlocks.setup_validate()
-        db.add_block(block1)
-        db.add_block(block2)
-        db.add_block(block3)
-        # Act
-        block2 = TrustChainBlock(block2.pack_db_insert())
-        block2.total_up += 10
-        block2.sign(db.get(block2.public_key, block2.sequence_number).key)
-        result = block2.validate(db)
-        # Assert
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn('Total up is higher than expected compared to the next block', result[1])
-
-    def test_validate_existing_total_down(self):
-        # Arrange
-        db = self.MockDatabase()
-        (block1, block2, block3, _) = TestBlocks.setup_validate()
-        db.add_block(block1)
-        db.add_block(block2)
-        db.add_block(block3)
-        # Act
-        block2 = TrustChainBlock(block2.pack_db_insert())
-        block2.total_down += 10
-        block2.sign(db.get(block2.public_key, block2.sequence_number).key)
-        result = block2.validate(db)
-        # Assert
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn('Total down is higher than expected compared to the next block', result[1])
-
     def test_validate_existing_link_public_key(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, block3, _) = TestBlocks.setup_validate()
         db.add_block(block1)
         db.add_block(block2)
@@ -317,7 +292,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_existing_link_sequence_number(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, block3, _) = TestBlocks.setup_validate()
         db.add_block(block1)
         db.add_block(block2)
@@ -333,7 +308,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_existing_hash(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, block3, _) = TestBlocks.setup_validate()
         db.add_block(block1)
         db.add_block(block2)
@@ -349,7 +324,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_existing_fraud(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, _, _) = TestBlocks.setup_validate()
         db.add_block(block1)
         db.add_block(block2)
@@ -364,7 +339,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_seq_not_genesis(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, _, _) = TestBlocks.setup_validate()
         # Act
         block1.previous_hash = sha256(str(random.randint(0, 100000))).digest()
@@ -376,7 +351,7 @@ class TestBlocks(TrustChainTestCase):
 
     def test_validate_seq_genesis(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, block3, _) = TestBlocks.setup_validate()
         db.add_block(block1)
         db.add_block(block3)
@@ -389,83 +364,9 @@ class TestBlocks(TrustChainTestCase):
         self.assertEqual(result[0], ValidationResult.invalid)
         self.assertIn('Sequence number implies previous hash should not be Genesis ID', result[1])
 
-    def test_validate_genesis(self):
-        # Arrange
-        db = self.MockDatabase()
-        (block1, _, _, _) = TestBlocks.setup_validate()
-        # Act
-        block1.up += 10
-        block1.down += 10
-        block1.sign(block1.key)
-        result = block1.validate(db)
-        # Assert
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn('Genesis block invalid total_up and/or up', result[1])
-        self.assertIn('Genesis block invalid total_down and/or down', result[1])
-
-    def test_validate_up(self):
-        # Arrange
-        db = self.MockDatabase()
-        (block1, block2, block3, _) = TestBlocks.setup_validate()
-        db.add_block(block1)
-        db.add_block(block3)
-        # Act
-        block2.up += 10
-        block2.sign(block2.key)
-        block3.previous_hash = block2.hash
-        result = block2.validate(db)
-        # Assert
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn('Total up is lower than expected compared to the preceding block', result[1])
-
-    def test_validate_down(self):
-        # Arrange
-        db = self.MockDatabase()
-        (block1, block2, block3, _) = TestBlocks.setup_validate()
-        db.add_block(block1)
-        db.add_block(block3)
-        # Act
-        block2.down += 10
-        block2.sign(block2.key)
-        block3.previous_hash = block2.hash
-        result = block2.validate(db)
-        # Assert
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn('Total down is lower than expected compared to the preceding block', result[1])
-
-    def test_validate_total_up(self):
-        # Arrange
-        db = self.MockDatabase()
-        (block1, block2, block3, _) = TestBlocks.setup_validate()
-        db.add_block(block1)
-        db.add_block(block3)
-        # Act
-        block2.total_up += 10
-        block2.sign(block2.key)
-        block3.previous_hash = block2.hash
-        result = block2.validate(db)
-        # Assert
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn('Total up is higher than expected compared to the next block', result[1])
-
-    def test_validate_total_down(self):
-        # Arrange
-        db = self.MockDatabase()
-        (block1, block2, block3, _) = TestBlocks.setup_validate()
-        db.add_block(block1)
-        db.add_block(block3)
-        # Act
-        block2.total_down += 10
-        block2.sign(block2.key)
-        block3.previous_hash = block2.hash
-        result = block2.validate(db)
-        # Assert
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn('Total down is higher than expected compared to the next block', result[1])
-
     def test_validate_hash(self):
         # Arrange
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, block3, _) = TestBlocks.setup_validate()
         db.add_block(block1)
         db.add_block(block3)
@@ -478,33 +379,8 @@ class TestBlocks(TrustChainTestCase):
         self.assertEqual(result[0], ValidationResult.invalid)
         self.assertIn('Previous hash is not equal to the hash id of the previous block', result[1])
 
-    def test_validate_not_sane_negatives(self):
-        db = self.MockDatabase()
-        block1 = TrustChainBlock()
-        # Act
-        block1.up = -10
-        block1.down = -20
-        block1.total_down = -10
-        block1.total_up = -20
-        result = block1.validate(db)
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn("Up field is negative", result[1])
-        self.assertIn("Down field is negative", result[1])
-        self.assertIn("Total up field is negative", result[1])
-        self.assertIn("Total down field is negative", result[1])
-
-    def test_validate_not_sane_zeroes(self):
-        db = self.MockDatabase()
-        block1 = TrustChainBlock()
-        # Act
-        block1.up = 0
-        block1.down = 0
-        result = block1.validate(db)
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn("Up and down are zero", result[1])
-
     def test_validate_not_sane_sequence_number(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, _, _) = TestBlocks.setup_validate()
         # Act
         block1.sequence_number = 0
@@ -514,7 +390,7 @@ class TestBlocks(TrustChainTestCase):
         self.assertIn("Sequence number is prior to genesis", result[1])
 
     def test_validate_not_sane_link_sequence_number(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, _, _) = TestBlocks.setup_validate()
         # Act
         block1.link_sequence_number = -1
@@ -523,7 +399,7 @@ class TestBlocks(TrustChainTestCase):
         self.assertIn("Link sequence number not empty and is prior to genesis", result[1])
 
     def test_validate_not_sane_public_key(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, _, _) = TestBlocks.setup_validate()
         # Act
         block1.public_key = EMPTY_PK
@@ -533,7 +409,7 @@ class TestBlocks(TrustChainTestCase):
         self.assertIn("Public key is not valid", result[1])
 
     def test_validate_not_sane_link_public_key(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, _, _) = TestBlocks.setup_validate()
         # Act
         block1.link_public_key = EMPTY_PK
@@ -543,7 +419,7 @@ class TestBlocks(TrustChainTestCase):
         self.assertIn("Linked public key is not valid", result[1])
 
     def test_validate_not_sane_self_signed(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, _, _) = TestBlocks.setup_validate()
         # Act
         block1.link_public_key = block1.public_key
@@ -553,7 +429,7 @@ class TestBlocks(TrustChainTestCase):
         self.assertIn("Self signed block", result[1])
 
     def test_validate_not_sane_invalid_signature(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, _, _) = TestBlocks.setup_validate()
         # Act
         block1.signature = EMPTY_SIG
@@ -562,53 +438,31 @@ class TestBlocks(TrustChainTestCase):
         self.assertIn("Invalid signature", result[1])
 
     def test_validate_linked_valid(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, _, _) = TestBlocks.setup_validate()
         db.add_block(block2)
         # Act
-        db.add_block(TrustChainBlock.create(db, block1.link_public_key, block1))
+        db.add_block(TrustChainBlock.create(block1.transaction, db, block1.link_public_key, block1))
         result = block1.validate(db)
         self.assertEqual(result[0], ValidationResult.valid)
 
-    def test_validate_linked_up(self):
-        db = self.MockDatabase()
-        (block1, block2, _, _) = TestBlocks.setup_validate()
-        db.add_block(block2)
-        # Act
-        db.add_block(TrustChainBlock.create(db, block1.link_public_key, block1))
-        block1.up += 5
-        result = block1.validate(db)
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn("Up/down mismatch on linked block", result[1])
-
-    def test_validate_linked_down(self):
-        db = self.MockDatabase()
-        (block1, block2, _, _) = TestBlocks.setup_validate()
-        db.add_block(block2)
-        # Act
-        db.add_block(TrustChainBlock.create(db, block1.link_public_key, block1))
-        block1.down -= 5
-        result = block1.validate(db)
-        self.assertEqual(result[0], ValidationResult.invalid)
-        self.assertIn("Down/up mismatch on linked block", result[1])
-
     def test_validate_linked_mismatch(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, block2, _, _) = TestBlocks.setup_validate()
         # Act
         db.add_block(block1)
-        block3 = TrustChainBlock.create(db, block2.link_public_key, block1)
+        block3 = TrustChainBlock.create(block2.transaction, db, block2.link_public_key, block1)
         result = block3.validate(db)
         self.assertEqual(result[0], ValidationResult.invalid)
         self.assertIn("Public key mismatch on linked block", result[1])
 
     def test_validate_linked_double_pay_fraud(self):
-        db = self.MockDatabase()
+        db = MockDatabase()
         (block1, _, _, _) = TestBlocks.setup_validate()
         # Act
         db.add_block(block1)
-        db.add_block(TrustChainBlock.create(db, block1.link_public_key, block1))
-        block2 = TrustChainBlock.create(db, block1.link_public_key, block1)
+        db.add_block(TrustChainBlock.create(block1.transaction, db, block1.link_public_key, block1))
+        block2 = TrustChainBlock.create(block1.transaction, db, block1.link_public_key, block1)
         result = block2.validate(db)
         self.assertEqual(result[0], ValidationResult.invalid)
         self.assertIn("Double countersign fraud", result[1])
@@ -619,8 +473,6 @@ class TestBlocks(TrustChainTestCase):
         block1 = TestBlock()
         block1.sequence_number = GENESIS_SEQ
         block1.previous_hash = GENESIS_HASH
-        block1.total_up = block1.up
-        block1.total_down = block1.down
         block1.sign(block1.key)
         block2 = TestBlock(previous=block1)
         block3 = TestBlock(previous=block2)
@@ -634,42 +486,3 @@ class TestBlocks(TrustChainTestCase):
         self.assertIsNone(ValidationResult.partial_next())
         self.assertIsNone(ValidationResult.partial_previous())
         self.assertIsNone(ValidationResult.valid())
-
-    class MockDatabase(object):
-        def __init__(self):
-            super(TestBlocks.MockDatabase, self).__init__()
-            self.data = dict()
-
-        def add_block(self, block):
-            if self.data.get(block.public_key) is None:
-                self.data[block.public_key] = []
-            self.data[block.public_key].append(block)
-            self.data[block.public_key].sort(key=lambda b: b.sequence_number)
-
-        def get(self, pk, seq):
-            if self.data.get(pk) is None:
-                return None
-            item = [i for i in self.data[pk] if i.sequence_number == seq]
-            return item[0] if item else None
-
-        def get_linked(self, blk):
-            if self.data.get(blk.link_public_key) is None:
-                return None
-            item = [i for i in self.data[blk.link_public_key] if
-                    i.sequence_number == blk.link_sequence_number or i.link_sequence_number == blk.sequence_number]
-            return item[0] if item else None
-
-        def get_latest(self, pk):
-            return self.data[pk][-1] if self.data.get(pk) else None
-
-        def get_block_after(self, blk):
-            if self.data.get(blk.public_key) is None:
-                return None
-            item = [i for i in self.data[blk.public_key] if i.sequence_number > blk.sequence_number]
-            return item[0] if item else None
-
-        def get_block_before(self, blk):
-            if self.data.get(blk.public_key) is None:
-                return None
-            item = [i for i in self.data[blk.public_key] if i.sequence_number < blk.sequence_number]
-            return item[-1] if item else None
