+++ Tribler/Test/Community/Trustchain/test_trustchain_utilities.py
@@ -4,6 +4,7 @@ File containing utilities used in testing the TrustChain community.
 import random
 from hashlib import sha256
 
+from Tribler.Core.Utilities.encoding import encode
 from Tribler.Test.test_as_server import AbstractServer
 from Tribler.dispersy.crypto import ECCrypto
 from Tribler.community.trustchain.block import TrustChainBlock
@@ -15,23 +16,20 @@ class TestBlock(TrustChainBlock):
     Also used in other test files for TrustChain.
     """
 
-    def __init__(self, previous=None):
+    def __init__(self, transaction=None, previous=None):
         crypto = ECCrypto()
-        up = random.randint(201, 220)
-        down = random.randint(221, 240)
         other = crypto.generate_key(u"curve25519").pub().key_to_bin()
+
+        transaction = transaction or {'id': 42}
 
         if previous:
             self.key = previous.key
-            TrustChainBlock.__init__(self, (
-                up, down, previous.total_up + up, previous.total_down + down,
-                previous.public_key, previous.sequence_number + 1, other, 0,
-                previous.hash, 0, 0))
+            TrustChainBlock.__init__(self, (encode(transaction), previous.public_key, previous.sequence_number + 1,
+                                            other, 0, previous.hash, 0, 0))
         else:
             self.key = crypto.generate_key(u"curve25519")
             TrustChainBlock.__init__(self, (
-                up, down, random.randint(241, 260), random.randint(261, 280),
-                self.key.pub().key_to_bin(), random.randint(50, 100), other, 0,
+                encode(transaction), self.key.pub().key_to_bin(), random.randint(50, 100), other, 0,
                 sha256(str(random.randint(0, 100000))).digest(), 0, 0))
         self.sign(self.key)
 
@@ -46,10 +44,7 @@ class TrustChainTestCase(AbstractServer):
         self.assertTrue(actual_block is not None)
         self.assertTrue(crypto.is_valid_public_bin(expected_block.public_key))
         self.assertTrue(crypto.is_valid_public_bin(actual_block.public_key))
-        self.assertEqual(expected_block.up, actual_block.up)
-        self.assertEqual(expected_block.down, actual_block.down)
-        self.assertEqual(expected_block.total_up, actual_block.total_up)
-        self.assertEqual(expected_block.total_down, actual_block.total_down)
+        self.assertDictEqual(expected_block.transaction, actual_block.transaction)
         self.assertEqual(expected_block.public_key, actual_block.public_key)
         self.assertEqual(expected_block.sequence_number, actual_block.sequence_number)
         self.assertEqual(expected_block.link_public_key, actual_block.link_public_key)
