+++ Tribler/community/triblerchain/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains code for the TriblerChain, specifically designed to account up- and download amounts.
+"""
