+++ Tribler/community/triblerchain/community.py
@@ -1,0 +1,159 @@
+from twisted.internet import reactor
+from twisted.internet.defer import inlineCallbacks
+from twisted.internet.task import LoopingCall
+
+from Tribler.Core.simpledefs import NTFY_TUNNEL, NTFY_REMOVE
+from Tribler.community.triblerchain.block import TriblerChainBlock
+from Tribler.community.triblerchain.database import TriblerChainDB
+from Tribler.community.trustchain.community import TrustChainCommunity
+from Tribler.dispersy.util import blocking_call_on_reactor_thread
+
+MIN_TRANSACTION_SIZE = 1024*1024
+
+
+class PendingBytes(object):
+    def __init__(self, up, down, clean=None):
+        super(PendingBytes, self).__init__()
+        self.up = up
+        self.down = down
+        self.clean = clean
+
+    def add(self, up, down):
+        if self.up + up >= 0 and self.down + down >= 0:
+            self.up = max(0, self.up + up)
+            self.down = max(0, self.down + down)
+            if self.clean is not None:
+                self.clean.reset(2 * 60)
+            return True
+        else:
+            return False
+
+
+class TriblerChainCommunity(TrustChainCommunity):
+    """
+    Community for reputation based on TrustChain tamper proof interaction history.
+    """
+    BLOCK_CLASS = TriblerChainBlock
+    DB_CLASS = TriblerChainDB
+    SIGN_DELAY = 5
+
+    def __init__(self, *args, **kwargs):
+        super(TriblerChainCommunity, self).__init__(*args, **kwargs)
+        self.notifier = None
+
+        # We store the bytes send and received in the tunnel community in a dictionary.
+        # The key is the public key of the peer being interacted with, the value a tuple of the up and down bytes
+        # This data is not used to create outgoing requests, but _only_ to verify incoming requests
+        self.pending_bytes = dict()
+
+    def initialize(self, tribler_session=None):
+        super(TriblerChainCommunity, self).initialize()
+        if tribler_session:
+            self.notifier = tribler_session.notifier
+            self.notifier.add_observer(self.on_tunnel_remove, NTFY_TUNNEL, [NTFY_REMOVE])
+
+    def should_sign(self, block):
+        """
+        Return whether we should sign the passed block.
+        @param block: the block that we should sign or not.
+        """
+        pend = self.pending_bytes.get(block.public_key)
+        if not pend or not (pend.up - block.transaction['down'] >= 0 and pend.down - block.transaction['up'] >= 0):
+            self.logger.info("Request block counter party does not have enough bytes pending.")
+            return False
+        return True
+
+    @blocking_call_on_reactor_thread
+    def get_statistics(self, public_key=None):
+        """
+        Returns a dictionary with some statistics regarding the local trustchain database
+        :returns a dictionary with statistics
+        """
+        if public_key is None:
+            public_key = self.my_member.public_key
+        latest_block = self.persistence.get_latest(public_key)
+        statistics = dict()
+        statistics["id"] = public_key.encode("hex")
+        interacts = self.persistence.get_num_unique_interactors(public_key)
+        statistics["peers_that_pk_helped"] = interacts[0] if interacts[0] is not None else 0
+        statistics["peers_that_helped_pk"] = interacts[1] if interacts[1] is not None else 0
+        if latest_block:
+            statistics["total_blocks"] = latest_block.sequence_number
+            statistics["total_up"] = latest_block.transaction["total_up"]
+            statistics["total_down"] = latest_block.transaction["total_down"]
+            statistics["latest_block"] = dict(latest_block)
+
+            # Set up/down
+            statistics["latest_block"]["up"] = latest_block.transaction["up"]
+            statistics["latest_block"]["down"] = latest_block.transaction["down"]
+        else:
+            statistics["total_blocks"] = 0
+            statistics["total_up"] = 0
+            statistics["total_down"] = 0
+        return statistics
+
+    @blocking_call_on_reactor_thread
+    def on_tunnel_remove(self, subject, change_type, tunnel, candidate):
+        """
+        Handler for the remove event of a tunnel. This function will attempt to create a block for the amounts that
+        were transferred using the tunnel.
+        :param subject: Category of the notifier event
+        :param change_type: Type of the notifier event
+        :param tunnel: The tunnel that was removed (closed)
+        :param candidate: The dispersy candidate with whom this node has interacted in the tunnel
+        """
+        from Tribler.community.tunnel.tunnel_community import Circuit, RelayRoute, TunnelExitSocket
+        assert isinstance(tunnel, Circuit) or isinstance(tunnel, RelayRoute) or isinstance(tunnel, TunnelExitSocket), \
+            "on_tunnel_remove() was called with an object that is not a Circuit, RelayRoute or TunnelExitSocket"
+        assert isinstance(tunnel.bytes_up, int) and isinstance(tunnel.bytes_down, int),\
+            "tunnel instance must provide byte counts in int"
+
+        up = tunnel.bytes_up
+        down = tunnel.bytes_down
+        pk = candidate.get_member().public_key
+
+        # If the transaction is not big enough we discard the bytes up and down.
+        if up + down >= MIN_TRANSACTION_SIZE:
+            # Tie breaker to prevent both parties from requesting
+            if up > down or (up == down and self.my_member.public_key > pk):
+                self.register_task("sign_%s" % tunnel.circuit_id,
+                                   reactor.callLater(self.SIGN_DELAY, self.sign_block, candidate, pk,
+                                                     {'up': tunnel.bytes_up, 'down': tunnel.bytes_down}))
+            else:
+                pend = self.pending_bytes.get(pk)
+                if not pend:
+                    self.pending_bytes[pk] = PendingBytes(up,
+                                                          down,
+                                                          reactor.callLater(2 * 60, self.cleanup_pending, pk))
+                else:
+                    pend.add(up, down)
+
+    def cleanup_pending(self, public_key):
+        self.pending_bytes.pop(public_key, None)
+
+    @inlineCallbacks
+    def unload_community(self):
+        if self.notifier:
+            self.notifier.remove_observer(self.on_tunnel_remove)
+        for pk in self.pending_bytes:
+            if self.pending_bytes[pk].clean is not None:
+                self.pending_bytes[pk].clean.reset(0)
+        yield super(TriblerChainCommunity, self).unload_community()
+
+
+class TriblerChainCommunityCrawler(TriblerChainCommunity):
+    """
+    Extended TriblerChainCommunity that also crawls other TriblerChainCommunities.
+    It requests the chains of other TrustChains.
+    """
+
+    # Time the crawler waits between crawling a new candidate.
+    CrawlerDelay = 5.0
+
+    def on_introduction_response(self, messages):
+        super(TriblerChainCommunityCrawler, self).on_introduction_response(messages)
+        for message in messages:
+            self.send_crawl_request(message.candidate, message.candidate.get_member().public_key)
+
+    def start_walking(self):
+        self.register_task("take step", LoopingCall(self.take_step)).start(self.CrawlerDelay, now=False)
