+++ Tribler/community/trustchain/community.py
@@ -6,12 +6,8 @@ Every node has a chain and these chains intertwine by blocks shared by chains.
 import logging
 from twisted.internet import reactor
 from twisted.internet.defer import inlineCallbacks
-from twisted.internet.task import LoopingCall
-
-from Tribler.Core.CacheDB.sqlitecachedb import forceDBThread
-from Tribler.Core.simpledefs import NTFY_TUNNEL, NTFY_REMOVE
-from Tribler.dispersy.authentication import MemberAuthentication
-from Tribler.dispersy.authentication import NoAuthentication
+from Tribler.community.trustchain.database import TrustChainDB
+from Tribler.dispersy.authentication import NoAuthentication, MemberAuthentication
 from Tribler.dispersy.community import Community
 from Tribler.dispersy.conversion import DefaultConversion
 from Tribler.dispersy.destination import CandidateDestination
@@ -19,62 +15,28 @@ from Tribler.dispersy.distribution import DirectDistribution
 from Tribler.dispersy.message import Message, DelayPacketByMissingMember
 from Tribler.dispersy.resolution import PublicResolution
 
-from Tribler.dispersy.util import blocking_call_on_reactor_thread
 from Tribler.community.trustchain.block import TrustChainBlock, ValidationResult, GENESIS_SEQ, UNKNOWN_SEQ
 from Tribler.community.trustchain.payload import HalfBlockPayload, CrawlRequestPayload
-from Tribler.community.trustchain.database import TrustChainDB
 from Tribler.community.trustchain.conversion import TrustChainConversion
 
 HALF_BLOCK = u"half_block"
 CRAWL = u"crawl"
-MIN_TRANSACTION_SIZE = 1024*1024
-
-
-class PendingBytes(object):
-    def __init__(self, up, down, clean=None):
-        super(PendingBytes, self).__init__()
-        self.up = up
-        self.down = down
-        self.clean = clean
-
-    def add(self, up, down):
-        if self.up + up >= 0 and self.down + down >= 0:
-            self.up = max(0, self.up + up)
-            self.down = max(0, self.down + down)
-            if self.clean is not None:
-                self.clean.reset(2 * 60)
-            return True
-        else:
-            return False
 
 
 class TrustChainCommunity(Community):
     """
     Community for reputation based on TrustChain tamper proof interaction history.
     """
+    BLOCK_CLASS = TrustChainBlock
     DB_CLASS = TrustChainDB
     DB_NAME = 'trustchain'
 
     def __init__(self, *args, **kwargs):
         super(TrustChainCommunity, self).__init__(*args, **kwargs)
         self.logger = logging.getLogger(self.__class__.__name__)
-
-        self.notifier = None
         self.persistence = self.DB_CLASS(self.dispersy.working_directory, self.DB_NAME)
-
-        # We store the bytes send and received in the tunnel community in a dictionary.
-        # The key is the public key of the peer being interacted with, the value a tuple of the up and down bytes
-        # This data is not used to create outgoing requests, but _only_ to verify incoming requests
-        self.pending_bytes = dict()
-
         self.logger.debug("The trustchain community started with Public Key: %s",
                           self.my_member.public_key.encode("hex"))
-
-    def initialize(self, tribler_session=None):
-        super(TrustChainCommunity, self).initialize()
-        if tribler_session:
-            self.notifier = tribler_session.notifier
-            self.notifier.add_observer(self.on_tunnel_remove, NTFY_TUNNEL, [NTFY_REMOVE])
 
     @classmethod
     def get_master_members(cls, dispersy):
@@ -124,6 +86,13 @@ class TrustChainCommunity(Community):
     def initiate_conversions(self):
         return [DefaultConversion(self), TrustChainConversion(self)]
 
+    def should_sign(self, block):
+        """
+        Return whether we should sign the passed block.
+        @param block: the block that we should sign or not.
+        """
+        return True
+
     def send_block(self, candidate, block):
         if candidate.get_member():
             self.logger.debug("Sending block to %s (%s)", candidate.get_member().public_key.encode("hex")[-8:], block)
@@ -139,31 +108,25 @@ class TrustChainCommunity(Community):
         except DelayPacketByMissingMember:
             self.logger.warn("Missing member in TrustChain community to send signature request to")
 
-    def sign_block(self, candidate, public_key=None, bytes_up=None, bytes_down=None, linked=None):
+    def sign_block(self, candidate, public_key=None, transaction=None, linked=None):
         """
         Create, sign, persist and send a block signed message
         :param candidate: The peer with whom you have interacted, as a dispersy candidate
-        :param bytes_up: The bytes you have uploaded to the peer in this interaction
-        :param bytes_down: The bytes you have downloaded from the peer in this interaction
+        :param transaction: A string describing the interaction in this block
         :param linked: The block that the requester is asking us to sign
         """
         # NOTE to the future: This method reads from the database, increments and then writes back. If in some future
         # this method is allowed to execute in parallel, be sure to lock from before .create up to after .add_block
-        assert bytes_up is None and bytes_down is None and linked is not None or \
-            bytes_up is not None and bytes_down is not None and linked is None, \
-            "Either provide a linked block or byte counts, not both"
+        assert transaction is None and linked is not None or transaction is not None and linked is None, \
+            "Either provide a linked block or a transaction, not both"
         assert linked is None or linked.link_public_key == self.my_member.public_key, \
             "Cannot counter sign block not addressed to self"
         assert linked is None or linked.link_sequence_number == UNKNOWN_SEQ, \
             "Cannot counter sign block that is not a request"
+        assert transaction is None or isinstance(transaction, dict), "Transaction should be a dictionary"
 
-        block = TrustChainBlock.create(self.persistence, self.my_member.public_key, linked)
-        if linked is None:
-            block.up = bytes_up
-            block.down = bytes_down
-            block.total_up += bytes_up
-            block.total_down += bytes_down
-            block.link_public_key = public_key
+        block = self.BLOCK_CLASS.create(transaction, self.persistence, self.my_member.public_key,
+                                        link=linked, link_pk=public_key)
         block.sign(self.my_member.private_key)
         validation = block.validate(self.persistence)
         self.logger.info("Signed block to %s (%s) validation result %s",
@@ -199,10 +162,8 @@ class TrustChainCommunity(Community):
 
             self.logger.info("Received request block addressed to us (%s)", blk)
 
-            # determine if we want to sign (i.e. the requesting public key has enough pending bytes)
-            pend = self.pending_bytes.get(blk.public_key)
-            if not pend or not pend.add(-blk.down, -blk.up):
-                self.logger.info("Request block counter party does not have enough bytes pending.")
+            # determine if we want to sign this block
+            if not self.should_sign(blk):
                 continue
 
             crawl_task = "crawl_%s" % blk.hash
@@ -214,9 +175,6 @@ class TrustChainCommunity(Community):
                 self.logger.info("Request block could not be validated sufficiently, crawling requester. %s",
                                  validation)
                 # Note that this code does not cover the scenario where we obtain this block indirectly.
-
-                # We modified the counters to get here, correct pending bytes since we did not really sign the block yet
-                pend.add(blk.down, blk.up)
 
                 # Are we already waiting for this crawl to happen?
                 # For example: it's taking longer than 5 secs or the block message reached us twice via different paths
@@ -260,96 +218,9 @@ class TrustChainCommunity(Community):
                 self.send_block(message.candidate, blk)
             self.logger.info("Sent %d blocks", count)
 
-    @blocking_call_on_reactor_thread
-    def get_statistics(self, public_key=None):
-        """
-        Returns a dictionary with some statistics regarding the local trustchain database
-        :returns a dictionary with statistics
-        """
-        if public_key is None:
-            public_key = self.my_member.public_key
-        latest_block = self.persistence.get_latest(public_key)
-        statistics = dict()
-        statistics["id"] = public_key.encode("hex")
-        interacts = self.persistence.get_num_unique_interactors(public_key)
-        statistics["peers_that_pk_helped"] = interacts[0] if interacts[0] is not None else 0
-        statistics["peers_that_helped_pk"] = interacts[1] if interacts[1] is not None else 0
-        if latest_block:
-            statistics["total_blocks"] = latest_block.sequence_number
-            statistics["total_up"] = latest_block.total_up
-            statistics["total_down"] = latest_block.total_down
-            statistics["latest_block"] = dict(latest_block)
-        else:
-            statistics["total_blocks"] = 0
-            statistics["total_up"] = 0
-            statistics["total_down"] = 0
-        return statistics
-
     @inlineCallbacks
     def unload_community(self):
         self.logger.debug("Unloading the TrustChain Community.")
-        if self.notifier:
-            self.notifier.remove_observer(self.on_tunnel_remove)
-        for pk in self.pending_bytes:
-            if self.pending_bytes[pk].clean is not None:
-                self.pending_bytes[pk].clean.reset(0)
         yield super(TrustChainCommunity, self).unload_community()
         # Close the persistence layer
         self.persistence.close()
-
-    @forceDBThread
-    def on_tunnel_remove(self, subject, change_type, tunnel, candidate):
-        """
-        Handler for the remove event of a tunnel. This function will attempt to create a block for the amounts that
-        were transferred using the tunnel.
-        :param subject: Category of the notifier event
-        :param change_type: Type of the notifier event
-        :param tunnel: The tunnel that was removed (closed)
-        :param candidate: The dispersy candidate with whom this node has interacted in the tunnel
-        """
-        from Tribler.community.tunnel.tunnel_community import Circuit, RelayRoute, TunnelExitSocket
-        assert isinstance(tunnel, Circuit) or isinstance(tunnel, RelayRoute) or isinstance(tunnel, TunnelExitSocket), \
-            "on_tunnel_remove() was called with an object that is not a Circuit, RelayRoute or TunnelExitSocket"
-        assert isinstance(tunnel.bytes_up, int) and isinstance(tunnel.bytes_down, int),\
-            "tunnel instance must provide byte counts in int"
-
-        up = tunnel.bytes_up
-        down = tunnel.bytes_down
-        pk = candidate.get_member().public_key
-
-        # If the transaction is not big enough we discard the bytes up and down.
-        if up + down >= MIN_TRANSACTION_SIZE:
-            # Tie breaker to prevent both parties from requesting
-            if up > down or (up == down and self.my_member.public_key > pk):
-                self.register_task("sign_%s" % tunnel.circuit_id,
-                                   reactor.callLater(5, self.sign_block, candidate, pk,
-                                                     tunnel.bytes_up, tunnel.bytes_down))
-            else:
-                pend = self.pending_bytes.get(pk)
-                if not pend:
-                    self.pending_bytes[pk] = PendingBytes(up,
-                                                          down,
-                                                          reactor.callLater(2 * 60, self.cleanup_pending, pk))
-                else:
-                    pend.add(up, down)
-
-    def cleanup_pending(self, public_key):
-        self.pending_bytes.pop(public_key, None)
-
-
-class TrustChainCommunityCrawler(TrustChainCommunity):
-    """
-    Extended TrustChainCommunity that also crawls other TrustChainCommunities.
-    It requests the chains of other TrustChains.
-    """
-
-    # Time the crawler waits between crawling a new candidate.
-    CrawlerDelay = 5.0
-
-    def on_introduction_response(self, messages):
-        super(TrustChainCommunityCrawler, self).on_introduction_response(messages)
-        for message in messages:
-            self.send_crawl_request(message.candidate, message.candidate.get_member().public_key)
-
-    def start_walking(self):
-        self.register_task("take step", LoopingCall(self.take_step)).start(self.CrawlerDelay, now=False)
